+++ server/src/typecheck/structs.ts
