--- server/src/vscode-parse.ts
@@ -1,9 +1,0 @@
-/**
- * This will fully parse one document,
- * using the given parser.
- * returning either a list of errors
- * or a list of declarations.
- * 
- * We need to use one parser because 
- * the lexer needs to be updated with typedefs each time
- */
